Neuer Ordner/requirements.txt
openai
anthropic
pyyaml
requests
<<<<<<< HEAD
streamlit
=======
streamlit
openai==0.28
>>>>>>> 131c861 (Buffixes // Fix OpenAI wrapper for openai v0.28.0 compatibility)
